Hover for explanation
I have my Eclipse workspace on OneDrive because I develop both on my
deskop computer and my Surface. I had to install the Java 1.8 JDK into a
folder on my OneDrive so the Eclipse programs on both computers can
reference the same JDK installation without needing to install it on
both separately. This will make replacement of one or both computers in
the future much easier.